my-project-server/requirements.txt
setproctitle>=1.1.10
itsdangerous>=0.24.0
Werkzeug>=0.12.1
flask>=0.12.1
MarkupSafe>=1.0
simplejson>=3.10.0
flask-socketio>=2.8.6
flask-admin>=1.5.0
flask-babelex>=0.9.3
flask-session>=0.3.1
gevent>=1.2.1
gevent-websocket>=0.10.1
future>=0.16.0
futures>=3.0.5 ; python_version < '3.2'
enum34>=1.1.6 ; python_version < '3.4'
configparser>=3.5.0 ; python_version < '3.2'
PyCRC>=1.21
pymongo>=3.4.0
mongoengine>=0.13.0
flask-mongoengine>=0.9.2
Pillow>=4.1.1
peewee>=2.10.1
wtf-peewee>=0.2.6
pymysql>=0.7.11